requirements.txt
alembic==1.7.6.dev0
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.12
asttokens==2.0.8
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
cloud-init==22.3.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dnspython==2.2.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
email-validator==1.3.0
entrypoints==0.4
executing==1.1.1
fastapi==0.85.1
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.6.0
flake8==5.0.4
future==0.18.2
greenlet==1.1.2
h11==0.14.0
httplib2==0.20.2
httptools==0.5.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.8.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.1
notebook_shim==0.2.0
oauthlib==3.2.0
olefile==0.46
orjson==3.8.1
packaging==21.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2022.10.25
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.2
pyflakes==2.5.0
Pygments==2.13.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylint==2.15.5
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-magic==0.4.24
python-multipart==0.0.5
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
Send2Trash==1.8.0
service-identity==18.1.0
six==1.16.0
sniffio==1.3.0
sos==4.4
soupsieve==2.3.2.post1
SQLAlchemy==1.4.31
ssh-import-id==5.11
stack-data==0.5.1
starlette==0.20.4
systemd-python==234
terminado==0.17.0
texttable==1.6.4
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.5.0
Twisted==22.1.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.5.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.18.3
uvloop==0.17.0
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
watchfiles==0.18.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.4
widgetsnbextension==4.0.3
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
